\name{dijkstra-package}
\alias{dijkstra-package}
\alias{dijkstra}
\docType{package}
\title{
\packageTitle{dijkstra}
}
\description{
\packageDescription{dijkstra}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{dijkstra}

Maintainer: \packageMaintainer{dijkstra}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = dijkstra)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{dijkstra}}
%\section{Documentation Index}{\packageIndices{dijkstra}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
