\name{dijkstra-package}
\alias{dijkstra-package}
\alias{dijkstra}
\docType{package}
\title{
\packageTitle{dijkstra}
}
\description{
\packageDescription{dijkstra}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dijkstra}
\packageIndices{dijkstra}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{dijkstra}

Maintainer: \packageMaintainer{dijkstra}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
